Set the kernel sys_tz variable via hwclock
This workaround uses a combo of ntpd to firstly set the date correctly, then
hwclock to set the kernel's timezone (`sys_tz` via `settimeofday()`).
With hwclock, the userspace system time is adjusted according to the TZ
offset (eg the clock is now wrong) so we finally start ntpd like normal
and let it correct itself thereafter.