Provide wpa_supplicant v2.7 dev to fix KRACK
This includes the compiled executables and libraries needed
to run wpa_supplicant with KRACK fixes on a Xiaofang camera.
The executables were compiled using the instructions in INSTALL.md and
can be deployed directly to a camera with firmware 3.0.3.56.  These have
not been tested on any other versions of the firmware.

This also fixes any other issues that came along in wpa_supplicant since
v0.8.x (which is the version in the stock firmware).

Take care when deploying these updated libraries and executables; they
may conflict with the camera's existing files so this is why I recommend
placing them into their own `updates` folder on the SD card (eg
`/media/mmcblk0p2/updates`).